feat: realtime 구독 훅 분리 및 사용자 ID 가져오기 로직 수정
