+++ src/components/layout/Header.tsx
@@ -3,17 +3,23 @@
 import { usePathname, useRouter } from "next/navigation";
 import PATH from "@/constants/PATH";
 import Image from "next/image";
-import { useState } from "react";
+import { useState,useEffect } from "react";
 import NotificationDropdown from "@/components/notifications/NotificationDropdown";
+import getUserData from "@/supabase/getUserData";
+import useNotificationSubscription from "@/hooks/useNotificationSubscription";
 import {
   DropdownMenu,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 
+
 const Header = () => {
   const router = useRouter();
   const pathname = usePathname();
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
+  const [newalarm, setNewalarm] = useState(false); // 새로운 알림 상태
+  const [userId, setUserId] = useState<string | null>(null);
+
 
   // 현재 랜딩 페이지인지 확인하기
   const isLandingPage = pathname === PATH.HOME;
@@ -36,6 +42,20 @@ const Header = () => {
   const handleHome = () => {
     router.push(PATH.HOME);
   };
+
+  useEffect(() => {
+    const fetchId = async () => {
+      const userData = await getUserData();
+      setUserId(userData?.user?.id || null);
+    };
+
+    fetchId();
+  }, []);
+
+  useNotificationSubscription(userId, (payload) => {
+    console.log("새로운 알림 도착 (헤더) - Hook!", payload.new);
+    setNewalarm(true);
+  });
 
   return (
     <header className="fixed top-0 z-50 flex w-full items-center justify-between border-b bg-white px-4 py-4">
@@ -68,14 +88,15 @@ const Header = () => {
         <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen}>
           <DropdownMenuTrigger asChild>
             <button className="relative ml-4">
+              {/* 조건부 이미지 렌더링 */}
               <Image
-                src="/icons/alarm.svg"
+                src={
+                  newalarm ? "/icons/alarm1.svg" : "/icons/alarm.svg"
+                }
                 alt="알림로고"
                 height={24}
                 width={24}
               />
-              
-              
             </button>
           </DropdownMenuTrigger>
           <NotificationDropdown
